Split the manager base class out into its own file
